requirements.txt
apiai==1.2.3
beautifulsoup4==4.6.0
certifi==2018.1.18
chardet==3.0.4
enum34==1.1.6
geolocation-python==0.2.2
geopy==1.12.0
get==0.0.39
idna==2.6
lxml==4.2.1
numpy==1.14.2
post==0.0.26
public==0.0.65
query-string==0.0.28
request==0.0.26
requests==2.18.4
urllib3==1.22
vk-api==10.1
wikipedia==1.4.0
feedparser==5.2.1
pymorphy2==0.8
kinopoiskpy==0.8.0
pyTelegramBotAPI==4.10.0
termcolor2~=0.0.3